requirements.txt
streamlit
pillow
pandas
requests
plotly
matplotlib
